add MongoDB, craete STUB-s for trait methods in case class implementations

+++ src/main/scala/route/handler/DeleteNoteHandlerLive.scala
@@ -1,6 +1,5 @@
 package route.handler
 
-import db.NotesRepository
 import route.implementation
 import route.interface._
 import route.implementation.DeleteNoteServiceLive
@@ -8,15 +7,16 @@ import zhttp.http.Response
 import zio.*
 import zhttp.http._
 import DeleteNoteHandlerLive.NoteID
+import db.note.NotesRepository
 
 
 trait DeleteNoteHandler {
-  def handle(noteId: NoteID, userId: Int): Task[Response]
+  def handle(noteId: NoteID, userId: Long): Task[Response]
 }
 
 final case class DeleteNoteHandlerLive(deleteNoteService: DeleteNoteService) extends DeleteNoteHandler {
 
-  override def handle(noteId: NoteID, userId: Int): Task[Response] = for {
+  override def handle(noteId: NoteID, userId: Long): Task[Response] = for {
     deleteStatus      <- deleteNoteService.deleteRecord(noteId, userId)
     response          <- ZIO.succeed(deleteStatus.fold(Response.text, Response.text))
   } yield response
