+++ src/main/scala/route/implementation/DeleteNoteServiceLive.scala
@@ -1,7 +1,8 @@
 package route.implementation
 
 import db.CRUD.DeletionStatus
-import db.{CRUD, NoteCRUD, NotesRepository}
+import db.note.{NoteCRUD, NotesRepository}
+import db.CRUD
 import model.Note
 import route.interface.DeleteNoteService
 import zhttp.http.Response
@@ -9,7 +10,7 @@ import zio.*
 final case class DeleteNoteServiceLive(notesRepository: NoteCRUD) extends DeleteNoteService {
 
 
-  override def deleteRecord(noteId: Int, userId: Int): Task[Either[String, String]] = 
+  override def deleteRecord(noteId: Long, userId: Long): Task[Either[String, String]] =
     notesRepository.deleteNoteByIdAndUserId(noteId, userId)
 
 }
