+++ src/main/scala/route/implementation/GetNoteServiceLive.scala
@@ -1,6 +1,7 @@
 package route.implementation
 
-import db.{CRUD, NoteCRUD, NotesRepository}
+import db.note.{NoteCRUD, NotesRepository}
+import db.CRUD
 import model.Note
 import route.interface.GetNoteService
 import zhttp.http.Response
@@ -9,7 +10,7 @@ import zio.json.*
 
 final case class GetNoteServiceLive(private final val notesRepository: NoteCRUD) extends GetNoteService {
 
-  override def getNote(noteId: Int, userId: Int): Task[Either[String, Note]] = for {
+  override def getNote(noteId: Long, userId: Long): Task[Either[String, Note]] = for {
     maybeNote <- notesRepository.getNoteByIdAndUserId(noteId, userId)
     response  <- ZIO.succeed(maybeNote.fold(Left(s"Could not find the note with id ${noteId.withQuotes}"))(Right(_)))
   } yield response
