+++ src/main/scala/route/interface/GetNoteService.scala
@@ -4,5 +4,5 @@ import model.Note
 import zio.*
 
 trait GetNoteService {
-  def getNote(noteId: Int, userId: Int): Task[Either[String, Note]]
+  def getNote(noteId: Long, userId: Long): Task[Either[String, Note]]
 }
